add oracle sdk.

+++ Jorm/Jorm.cpp
@@ -78,8 +78,8 @@ int _tmain(int argc, _TCHAR* argv[])
 
 	vector<string> keys = rs.GetAllKeys();
 
-	int len = keys.size();
-	for (int i = 0; i < len; i++) {
+	size_t len = keys.size();
+	for (size_t i = 0; i < len; i++) {
 		string v, key = keys[i];
 		int vType;
 		rs.GetValueAndTypeByKey(key, &v, &vType);
